Initial netlify configuration

+++ build.sh
@@ -1,6 +1,6 @@
+[ -e posts ] && rm -r posts
 pipenv run python get_posts.py
-rm -r blog/content/blog
+[ -e blog/content/blog ] && rm -r blog/content/blog
 cp -r posts blog/content/blog
 cd blog
 yarn build
-netlify deploy
